% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_universe_pkgs_json.R
\name{create_universe_pkgs_json}
\alias{create_universe_pkgs_json}
\title{Create Universe Packages JSON}
\usage{
create_universe_pkgs_json(
  group = "KWB-R",
  ignore_pkgs = NULL,
  non_r_packages = kwb.pkgstatus::get_non_r_packages(),
  github_token = get_github_token()
)
}
\arguments{
\item{group}{default: "KWB-R"}

\item{ignore_pkgs}{vector of pkg names that should be ignored to prevent
problems with codemeta generation (e.g. "kwb.read" fails due to missing
metadata on the license!)}

\item{non_r_packages}{vector with repos that are not R packages (default:
kwb.pkgstatus::get_non_r_packages())}

\item{github_token}{optionally a Github token in order to access private
repositories. Default: pkgmeta:::get_github_token()}
}
\value{
JSON structure as required for packages.json
}
\description{
Prepares a packages.json file required for R-universe for creating
our personal CRAN-like repository on R-universe
}
\examples{
create_universe_pkgs_json()


}
\references{
\url{https://ropensci.org/blog/2021/06/22/setup-runiverse/#the-packagesjson-registry-file}
}
