% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_gh_ratelimit.R
\name{get_gh_ratelimit}
\alias{get_gh_ratelimit}
\title{Get GitHub rate limit}
\usage{
get_gh_ratelimit(github_token = get_github_token())
}
\arguments{
\item{github_token}{token passed to \link[gh]{gh}.
Default: pkgmeta:::get_github_token()}
}
\value{
overview of rate limit
}
\description{
Get GitHub rate limit
}
\examples{
get_gh_ratelimit()
}
