% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/github_commits.R
\name{get_github_commits}
\alias{get_github_commits}
\title{Get Github Commits for One Repo}
\usage{
get_github_commits(repo, github_token = Sys.getenv("GITHUB_PAT"))
}
\arguments{
\item{repo}{full name of Github repo ("owner/repo_name", e.g. "kwb-r/kwb.utils")}

\item{github_token}{default: Sys.getenv("GITHUB_PAT")}
}
\value{
data frame
}
\description{
Get Github Commits for One Repo
}
\examples{
\dontrun{
pkg_commits <- pkgmeta::get_github_commits("kwb-r/kwb.utils")
head(pkg_commits )
}
}
