% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_github_packages.R
\name{get_github_packages}
\alias{get_github_packages}
\title{Get Github R Packages}
\usage{
get_github_packages(group = "KWB-R",
  github_token = getOption("github_token"))
}
\arguments{
\item{group}{default: "KWB-R"}

\item{github_token}{optionally a Github token in order to access private
repositories (default: NULL)}
}
\value{
data frame with R packages on Github
}
\description{
Get Github R Packages
}
\examples{
pkgs <- get_github_packages()
head(pkgs)

}
