% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_github_repos_json.R
\name{get_github_topics}
\alias{get_github_topics}
\title{Helper function: Get Github Topics}
\usage{
get_github_topics(full_names = get_github_full_names())
}
\arguments{
\item{full_names}{vector with full_names to Github repo e.g.
c("kwb-r/kwb.utils", "kwb-r/kwb.base"), default: get_github_full_names()}
}
\description{
Helper function: Get Github Topics
}
