% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_meta.R
\name{get_meta}
\alias{get_meta}
\title{Return meta data on CRAN and github packages}
\usage{
get_meta(
  cran_author = NULL,
  include_downloads = FALSE,
  start = "2000-01-01",
  github_repos = NULL,
  prefer_cran = TRUE
)
}
\arguments{
\item{cran_author}{Character string to find in CRAN author names}

\item{include_downloads}{Should total CRAN downloads since \code{start} be added to the tibble?}

\item{start}{Start date for download statistics. Ignored if \code{include_downloads} is \code{FALSE}.}

\item{github_repos}{Character vector of github repos}

\item{prefer_cran}{When a package is both on CRAN and github, which information should be preferenced?}
}
\description{
Find meta data for all packages on CRAN that match a given author and all packages contained in
a vector of github repositories and return as a tibble. Meta data on packages that are
both on CRAN and github will be combined into a single row. Where meta data clash, the
CRAN version is preferred when \code{prefer_cran} is \code{TRUE}, otherwise the github
version is preferred.
}
\examples{
get_meta(cran_author = "Hyndman")
get_meta(github_repos = c("robjhyndman/forecast", "earowang/hts"))
get_meta(cran_author = "Emi Tanaka", github_repos = c("numbats/yowie", "numbats/monash"))
}
