% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_cran.R
\name{get_meta_cran}
\alias{get_meta_cran}
\title{Return meta data on CRAN packages}
\usage{
get_meta_cran(packages, include_downloads = FALSE, start = "2000-01-01")
}
\arguments{
\item{packages}{Character vector of CRAN package names}

\item{include_downloads}{Should total CRAN downloads since \code{start} be added to the tibble?}

\item{start}{Start date for download statistics. Ignored if \code{include_downloads} is \code{FALSE}.}

\item{github_repos}{Character string of github repos}

\item{prefer_cran}{When a package is both on CRAN and github, which information should be preferenced?}
}
\description{
Find meta data for a vector of CRAN packages and return as a tibble.
}
\examples{
get_meta(cran_author = "Hyndman")
get_meta(github_repos = c("robjhyndman/forecast", "earowang/hts"))
get_meta(cran_author = "Emi Tanaka", github_repos = c("numbats/yowie", "numbats/monash"))
}
