% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_r_releases.R
\name{get_r_releases}
\alias{get_r_releases}
\title{Get R Releases}
\usage{
get_r_releases()
}
\value{
data frame with R releases
}
\description{
Get R Releases
}
\examples{
releases <- get_r_releases()
head(releases)
}
