% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/github_package_versions.R
\name{github_package_versions}
\alias{github_package_versions}
\title{Get Github Versions for One Repo}
\usage{
github_package_versions(repo, github_token = get_github_token())
}
\arguments{
\item{repo}{full name of Github repo ("owner/repo_name", e.g. "kwb-r/kwb.utils")}

\item{github_token}{Default: pkgmeta:::get_github_token()}
}
\value{
data frame
}
\description{
Get Github Versions for One Repo
}
\examples{
\dontrun{
pkg_versions <- pkgmeta::github_package_versions("kwb-r/kwb.utils")
head(pkg_versions)
}
}
