% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_kwb_github_packages.R
\name{install_kwb_github_packages}
\alias{install_kwb_github_packages}
\title{Install KWB Github Packages}
\usage{
install_kwb_github_packages(pkgs_kwb, dependencies = TRUE, quiet = TRUE, ...)
}
\arguments{
\item{pkgs_kwb}{vector with R packages to be installed}

\item{dependencies}{dependencies (default: TRUE)}

\item{quiet}{(default: TRUE)}

\item{...}{additional arguments passed to \link[remotes]{install_github}}
}
\value{
install R packages
}
\description{
Install KWB Github Packages
}
