% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myRpackages-package.R
\docType{package}
\name{myRpackages-package}
\alias{myRpackages}
\alias{myRpackages-package}
\title{myRpackages: Find details about my R packages}
\description{
Construct a tibble containing information about R packages with specified authors.
}
\author{
\strong{Maintainer}: Rob Hyndman \email{Rob.Hyndman@monash.com} (\href{https://orcid.org/0000-0002-2140-5352}{ORCID})

}
\keyword{internal}
