% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myRpackages-package.R
\docType{package}
\name{pkgmeta-package}
\alias{pkgmeta}
\alias{pkgmeta-package}
\title{pkgmeta: Find details about my R packages}
\description{
Finds meta data for all packages on CRAN that match a given author and packages contained in specified github repositories.
}
\author{
\strong{Maintainer}: Rob Hyndman \email{Rob.Hyndman@monash.com} (\href{https://orcid.org/0000-0002-2140-5352}{ORCID})

}
\keyword{internal}
