% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_commits_github.R
\name{plot_commits_github}
\alias{plot_commits_github}
\title{Plot Commits GitHub}
\usage{
plot_commits_github(repos_commits)
}
\arguments{
\item{repos_commits}{tibble as retrieved by
\code{\link{github_commits_repos}}}
}
\value{
ggplot of temporal development of commits for KWB-R on Github
}
\description{
Plot Commits GitHub
}
\examples{
\dontrun{
repos <- kwb.pkgstatus::get_github_repos(github_token = Sys.getenv("GITHUB_PAT")
repos_commits <- pkgmeta::github_commits_repos(repos$full_name)
pkgmeta::plot_commits_github(repos_commits)
}

}
