% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_rpackages_releases.R
\name{plot_github_pkgs_releases}
\alias{plot_github_pkgs_releases}
\title{Plot Github Package Releases Over Time}
\usage{
plot_github_pkgs_releases(pkgs_releases, last_update = Sys.time())
}
\arguments{
\item{pkgs_releases}{data frame with pkgs as retrieved by
\code{\link{github_packages_versions}}}

\item{last_update}{default: Sys.time()}
}
\value{
ggplot of temporal development of R package releases on Github
}
\description{
Plot Github Package Releases Over Time
}
\examples{
\dontrun{
pkgs <- get_github_packages()
pkgs_releases <- github_packages_versions(repos = pkgs$full_name)
plot_github_pkgs_releases(pkgs_releases)
}

}
