% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_r_releases.R
\name{plot_r_releases}
\alias{plot_r_releases}
\title{Plot R Releases}
\usage{
plot_r_releases(r_releases = get_r_releases())
}
\arguments{
\item{r_releases}{as retrieved by \link{get_r_releases}, (default: get_r_releases())}
}
\value{
plotly with R releases
}
\description{
Plot R Releases
}
\examples{
\dontrun{
plot_r_releases()
}
}
