% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/travis_pkg_install_script.R
\name{travis_pkg_install_script}
\alias{travis_pkg_install_script}
\title{Travis Package Install Script}
\usage{
travis_pkg_install_script(pkgs = pkgmeta::get_github_packages())
}
\arguments{
\item{pkgs}{datadframe with R packages as retrieved by
get_github_packages()}
}
\value{
installs kwb-r packages
}
\description{
Needs to be Run before rendering vignettes
}
