% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_pkg_github_json.R
\name{write_github_repos_json}
\alias{write_github_repos_json}
\title{Write Github Metadata to JSON}
\usage{
write_github_repos_json(github_repos = get_github_repos(),
  file = file.path(getwd(), "github.json"), dbg = TRUE)
}
\arguments{
\item{file}{path where to save github.json (default: file.path(getwd(),
"github.json"))}

\item{dbg}{should debug messages be printed? (default: TRUE)}

\item{pkgs}{data frame with R packages (default: get_github_packages())}
}
\value{
writes "github.json"
}
\description{
Write Github Metadata to JSON
}
