% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_pkg_codemeta_json.R
\name{write_pkg_codemeta_json}
\alias{write_pkg_codemeta_json}
\title{Write codemeta}
\usage{
write_pkg_codemeta_json(
  codemeta = create_pkg_codemeta(),
  file = file.path(getwd(), "codemetar.json"),
  dbg = TRUE
)
}
\arguments{
\item{codemeta}{codemeta object (as retrieved by create_codemeta()),
default: create_pkg_codemeta())}

\item{file}{path where to save codemeta.json (default: file.path(getwd(),
"codemetar.json"))}

\item{dbg}{should debug messages be printed? (default: TRUE)}
}
\value{
writes "codemeta.json"
}
\description{
Write codemeta
}
