% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_gh_ratelimit.R
\name{get_gh_ratelimit}
\alias{get_gh_ratelimit}
\title{Get GitHub rate limit}
\usage{
get_gh_ratelimit(github_token = Sys.getenv("GITHUB_PAT"))
}
\arguments{
\item{github_token}{token passed to \link[gh]{gh} (default:
Sys.getenv("GITHUB_PAT")}
}
\value{
overview of rate limit
}
\description{
Get GitHub rate limit
}
\examples{
get_gh_ratelimit()
}
