% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/github_commits.R
\name{get_github_commits_repos}
\alias{get_github_commits_repos}
\title{Get Github Commits for Multiple Repos}
\usage{
get_github_commits_repos(repos, github_token = Sys.getenv("GITHUB_PAT"))
}
\arguments{
\item{repos}{vector with full names of Github repos ("owner/repo_name",
e.g. c("kwb-r/kwb.utils", "kwb-r/kwb.ml", "kwb-r/aquanes.report"))}

\item{github_token}{default: Sys.getenv("GITHUB_PAT")}
}
\value{
data frame for all repos with releases
}
\description{
Get Github Commits for Multiple Repos
}
\examples{
\dontrun{
#repos <- kwb.pkgstatus::get_github_repos(github_token = Sys.getenv("GITHUB_PAT"))
#repos <- repos$full_name
repos <- paste0("kwb-r/", c("aquanes.report", "kwb.ml", "kwb.utils"))
pkgs_commits <- pkgmeta::get_github_commits_repos(repos)
head(pkgs_commits)
}
}
