% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_github_packages.R
\name{get_github_packages}
\alias{get_github_packages}
\title{Get Github R Packages}
\usage{
get_github_packages(group = "KWB-R", ignore_pkgs = NULL,
  non_r_packages = kwb.pkgstatus::get_non_r_packages(),
  github_token = getOption("github_token"))
}
\arguments{
\item{group}{default: "KWB-R"}

\item{ignore_pkgs}{vector of pkg names that should be ignored to prevent
problems with codemeta generation (e.g. "kwb.read" fails due to missing
metadata on the license!)}

\item{non_r_packages}{vector with repos that are not R packages (default:
kwb.pkgstatus::get_non_r_packages())}

\item{github_token}{optionally a Github token in order to access private
repositories (default: getOption("github_token"))}
}
\value{
data frame with R packages on Github
}
\description{
Get Github R Packages
}
\examples{
\dontrun{
pkgs <- get_github_packages()
head(pkgs)
}

}
