% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_github_repos_json.R
\name{get_github_repos}
\alias{get_github_repos}
\title{Helper function: Get Github Metadata For Repos}
\usage{
get_github_repos(group = "KWB-R",
  github_token = getOption("github_token"), dbg = TRUE, ...)
}
\arguments{
\item{group}{organisation (default: "KWB-R")}

\item{github_token}{default: getOption("github_token") or Sys.getenv("GITHUB_PAT")}

\item{dbg}{should debug messages be printed? (default: TRUE)}

\item{...}{additional arguments passed to gh:::gh_build_request(), see
https://developer.github.com/v3/ (e.g. type = "public")}
}
\description{
Helper function: Get Github Metadata For Repos
}
