% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkgmeta-package.R
\docType{package}
\name{pkgmeta-package}
\alias{pkgmeta}
\alias{pkgmeta-package}
\title{pkgmeta: Find details about R packages on CRAN or github}
\description{
Finds meta data for all packages on CRAN that match a given author and packages contained in specified github repositories.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/robjhyndman/pkgmeta/}
  \item \url{https://pkg.robjhyndman.com/pkgmeta/}
  \item Report bugs at \url{https://github.com/robjhyndman/pkgmeta/issues}
}

}
\author{
\strong{Maintainer}: Rob Hyndman \email{Rob.Hyndman@monash.com} (\href{https://orcid.org/0000-0002-2140-5352}{ORCID})

}
\keyword{internal}
